requirements.txt
flask
numpy
pandas
langdetect
pymorphy2
requests
nltk
psycopg2-binary
elasticsearch
python-dotenv
scikit-learn